backend/data-manager/requirements.txt
google-cloud-storage
google-cloud-pubsub
python-multipart
google-cloud-aiplatform